moved styles
